Пункт с перезагрузкой окна деталей выполнен. Удалены предупреждения в консоли.
