+++ src/components/protected-route/index.jsx
@@ -1,4 +1,4 @@
-import { useEffect } from "react";
+import { useCallback, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useNavigate } from "react-router";
 import { getProfile } from "../../services/middleware/auth";
@@ -10,13 +10,13 @@ const ProtectedRoute = (props) => {
 
     const reduxDispatch = useDispatch();
 
-    const init = () => {
+    const init = useCallback(() => {
         reduxDispatch(getProfile());
-    };
+    }, [reduxDispatch]);
 
     useEffect(() => {
         init();
-    }, []);
+    }, [init]);
 
     const navigate = useNavigate();
 
@@ -25,7 +25,7 @@ const ProtectedRoute = (props) => {
             navigate("/login");
         }
         
-    }, [authed, loaded]);
+    }, [authed, loaded, navigate]);
 
     return (
         <>
