+++ src/pages/logout.jsx
@@ -1,20 +1,16 @@
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { useNavigate } from "react-router";
 import { logout } from "../services/middleware/auth";
 
 const LogoutPage = () => {
     const reduxDispatch = useDispatch();
-    const navigate = useNavigate();
     const logoutSuccess = useSelector(store => store.authReducer.logoutSuccess);
 
     useEffect(() => {
-        console.log("logoutstart");
         reduxDispatch(logout());
-    }, []);
+    }, [reduxDispatch]);
 
     useEffect(() => {
-        console.log("LOGOUT", logoutSuccess);
         if (logoutSuccess){
             // navigate("/login", {state: {logoutPassed: true}});
         }
