+++ src/services/cookies.js
@@ -1,0 +1,32 @@
+export function deleteCookie(name) {
+  setCookie(name, null, { expires: -1 });
+}
+
+export function setCookie(name, value, props) {
+    props = props || {};
+    let exp = props.expires;
+    if (typeof exp == 'number' && exp) {
+      const d = new Date();
+      d.setTime(d.getTime() + exp * 1000);
+      exp = props.expires = d;
+    }
+    if (exp && exp.toUTCString) {
+      props.expires = exp.toUTCString();
+    }
+    value = encodeURIComponent(value);
+    let updatedCookie = name + '=' + value;
+    for (const propName in props) {
+      updatedCookie += '; ' + propName;
+      const propValue = props[propName];
+      if (propValue !== true) {
+        updatedCookie += '=' + propValue;
+      }
+    }
+    document.cookie = updatedCookie;
+}
+
+export function getCookie(name) {
+    // eslint-disable-next-line 
+    const matches = document.cookie.match(new RegExp('(?:^|; )' + name.replace(/([\.$?*|{}\(\)\[\]\\\/\+^])/g, '\\$1') + '=([^;]*)'));
+    return matches ? decodeURIComponent(matches[1]) : undefined;
+}
