+++ src/services/middleware/auth.js
@@ -1,41 +1,12 @@
 import { LOGIN, LOGOUT, REGISTER, RESET_PASSWORD, SAVE_PASSWORD, UPDATE_PROFILE, LOAD_PROFILE_FAILED } from '../actions/auth';
+import { deleteCookie, getCookie, setCookie } from '../cookies';
 
 const ENDPOINT_URL = 'https://norma.nomoreparties.space/api/';
 const LOGIN_URL = ENDPOINT_URL + 'auth/login';
 const LOGOUT_URL = ENDPOINT_URL +  'auth/logout';
 const REGISTER_URL = ENDPOINT_URL + 'auth/register';
-// const TOKEN_URL = ENDPOINT_URL + 'auth/token';
 const RESET_PASSWORD_STEP_1_URL = ENDPOINT_URL + 'password-reset';
 const RESET_PASSWORD_STEP_2_URL = ENDPOINT_URL + 'password-reset/reset';
-
-function setCookie(name, value, props) {
-    props = props || {};
-    let exp = props.expires;
-    if (typeof exp == 'number' && exp) {
-      const d = new Date();
-      d.setTime(d.getTime() + exp * 1000);
-      exp = props.expires = d;
-    }
-    if (exp && exp.toUTCString) {
-      props.expires = exp.toUTCString();
-    }
-    value = encodeURIComponent(value);
-    let updatedCookie = name + '=' + value;
-    for (const propName in props) {
-      updatedCookie += '; ' + propName;
-      const propValue = props[propName];
-      if (propValue !== true) {
-        updatedCookie += '=' + propValue;
-      }
-    }
-    document.cookie = updatedCookie;
-}
-
-export function getCookie(name) {
-    // eslint-disable-next-line 
-    const matches = document.cookie.match(new RegExp('(?:^|; )' + name.replace(/([\.$?*|{}\(\)\[\]\\\/\+^])/g, '\\$1') + '=([^;]*)'));
-    return matches ? decodeURIComponent(matches[1]) : undefined;
-}
 
 const resetPasswordInternal = async (url, email, dispatch) => {
     const data = { email };
@@ -153,7 +124,7 @@ const logoutInternal = async (url, dispatch) => {
         console.log("LOGIT INTERN", data.success);
         if (data.success){
             console.log(data.message);
-            localStorage.removeItem("accessToken");
+            deleteCookie("accessToken");
             localStorage.removeItem("refreshToken");
             dispatch({ type: LOGOUT });
         } else {
@@ -187,8 +158,7 @@ const loginInternal = async (url, email, password, dispatch) => {
     if (response.ok) {
         const data = await response.json();
         if (data.success){
-            localStorage.setItem("accessToken", data.accessToken);
-            // setCookie("accessToken", data.accessToken);
+            setCookie("accessToken", data.accessToken);
             localStorage.setItem("refreshToken", data.refreshToken);
             dispatch({ type: LOGIN, data });
         } else {
@@ -209,7 +179,7 @@ export const login = (email, password) => async (dispatch)  => {
 
 const updateProfileInternal = async (url, name, password, email, dispatch) => {
     const data = { name, password, email };
-    const accessToken = localStorage.getItem("accessToken");
+    const accessToken = getCookie("accessToken");
 
     const response = await fetch(url, {
         method: 'PATCH',
@@ -244,7 +214,7 @@ export const updateProfile = (name, password, email) => async (dispatch)  => {
 }
 
 const getProfileInternal = async (url, dispatch) => {
-    const accessToken = localStorage.getItem("accessToken");
+    const accessToken = getCookie("accessToken");
     
     const response = await fetch(url, {
         method: 'GET',
